+++ src/routes/ready.ts
@@ -1,0 +1,3 @@
+export function readyHandler() {
+  console.log("Bot is online!");
+}
